+++ packages/cli/templates/rust/quickgui.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "@quickgui/cli";
 
 export default defineConfig({
-  language: "go",
+  language: "rust",
   name: {{APP_NAME}},
   identifier: {{IDENTIFIER}},
   entry: ".",